+++ sdk/spring/azure-spring-boot-test-aad-resource-server/src/test/java/com/azure/test/aad/resource/server/AADWeiResourceServerIT.java
@@ -3,22 +3,24 @@
 
 package com.azure.test.aad.resource.server;
 
+import com.azure.spring.aad.webapi.AADJwtBearerTokenAuthenticationConverter;
 import com.azure.spring.aad.webapi.AADResourceServerWebSecurityConfigurerAdapter;
 import com.azure.spring.test.aad.AADWebApiITHelper;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Configuration;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.client.HttpClientErrorException;
 
+import java.security.Principal;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
@@ -26,8 +28,8 @@ import java.util.Map;
 import static com.azure.spring.test.Constant.MULTI_TENANT_SCOPE_GRAPH_READ;
 import static com.azure.spring.test.EnvironmentVariable.AAD_MULTI_TENANT_CLIENT_ID;
 import static com.azure.spring.test.EnvironmentVariable.AAD_MULTI_TENANT_CLIENT_SECRET;
+import static com.azure.spring.test.EnvironmentVariable.AAD_USER_NAME_1;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 public class AADWeiResourceServerIT {
@@ -35,7 +37,7 @@ public class AADWeiResourceServerIT {
     private AADWebApiITHelper aadWebApiITHelper;
 
     @BeforeAll
-    public void init() {
+    public void beforeAll() {
         Map<String, String> properties = new HashMap<>();
         properties.put("azure.activedirectory.client-id", AAD_MULTI_TENANT_CLIENT_ID);
         properties.put("azure.activedirectory.client-secret", AAD_MULTI_TENANT_CLIENT_SECRET);
@@ -55,22 +57,29 @@ public class AADWeiResourceServerIT {
 
     @Test
     public void testHasNoScope() {
-        assertThrows(HttpClientErrorException.class, () -> aadWebApiITHelper.httpGetStringByAccessToken("notExist"));
+        Assertions.assertThrows(HttpClientErrorException.class,
+            () -> aadWebApiITHelper.httpGetStringByAccessToken("notExist"));
+    }
+
+    @Test
+    public void testPrincipalName() {
+        assertEquals(AAD_USER_NAME_1, aadWebApiITHelper.httpGetStringByAccessToken("principalName"));
     }
 
     @EnableWebSecurity
     @EnableGlobalMethodSecurity(securedEnabled = true, prePostEnabled = true)
     @SpringBootApplication
     @RestController
-    public static class TestApp extends WebSecurityConfigurerAdapter {
+    public static class TestApp extends AADResourceServerWebSecurityConfigurerAdapter {
 
-        @Configuration
-        public static class DefaultAADResourceServerWebSecurityConfigurerAdapter extends
-            AADResourceServerWebSecurityConfigurerAdapter {
-            @Override
-            protected void configure(HttpSecurity http) throws Exception {
-                super.configure(http);
-            }
+        @Override
+        protected void configure(HttpSecurity http) throws Exception {
+            super.configure(http);
+            AADJwtBearerTokenAuthenticationConverter converter = new AADJwtBearerTokenAuthenticationConverter();
+            converter.setPrincipalClaimName("upn");
+            http.oauth2ResourceServer()
+                .jwt()
+                .jwtAuthenticationConverter(converter);
         }
 
         @GetMapping("graph")
@@ -83,6 +92,11 @@ public class AADWeiResourceServerIT {
         @PreAuthorize("hasAuthority('SCOPE_NotExist')")
         public String notExist() {
             return "notExist";
+        }
+
+        @GetMapping(value = "/principalName")
+        public ResponseEntity<String> principalName(Principal principal) {
+            return ResponseEntity.ok(principal.getName());
         }
     }
 }