+++ sdk/keyvault/azure-security-keyvault-jca/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Release History
 
 ## 1.0.0-beta.6 (Unreleased)
-
+### Breaking Changes
+ - Remove configurable property of azure.keyvault.aad-authentication-url which is configured according to azure.keyvault.uri automatically [#20530](https://github.com/Azure/azure-sdk-for-java/pull/20530)
 
 ## 1.0.0-beta.5 (2021-03-22)