+++ sdk/spring/azure-spring-boot-starter-keyvault-certificates/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release History
 
 ## 3.0.0-beta.6 (Unreleased)
-
+- Remove configurable property of azure.keyvault.aad-authentication-url which is configured according to azure.keyvault.uri automatically [#20530](https://github.com/Azure/azure-sdk-for-java/pull/20530)
 
 ## 3.0.0-beta.5 (2021-03-22)
 ### New Features